Move bulk of "golden tests" logic to libs/testdiff
This makes it available to non-integration tests as well.
